#57: divide ProcessGuid class to a separate file.
