+++ CreateGUIDVSPlugin/CreateGUIDVSPlugin.csproj
@@ -72,6 +72,7 @@
     <Compile Include="RenewGuidCurrentCommand.cs" />
     <Compile Include="Utility\Configuration.cs" />
     <Compile Include="Utility\FormatGuid.cs" />
+    <Compile Include="Utility\ProcessGuid.cs" />
     <Compile Include="Utility\ProcessTemplate.cs" />
     <Compile Include="Utility\ReplaceWithNewGuid.cs" />
     <Compile Include="Utility\Template.cs" />
